Moved ingredientSearchView into its own folder and made files for its subviews
